processor/requirements.txt
google-generativeai
pydantic
pydantic-settings
pyyaml
python-dotenv
python-dateutil
httpx
pytest
